Move tests to a subfolder (will add RequestTests) + import Throwable
